More consistency in the choice of native integer types for the (very old) census code.

+++ engine/census/gluingperms.h
@@ -239,7 +239,7 @@ class GluingPerms : public Output<GluingPerms<dim>> {
          *
          * @return the number of simplices under consideration.
          */
-        unsigned size() const;
+        size_t size() const;
 
         /**
          * Returns the specific pairing of simplex facets that this
@@ -299,7 +299,7 @@ class GluingPerms : public Output<GluingPerms<dim>> {
          * investigation (between 0 and \a dim inclusive).
          * @return the associated gluing permutation.
          */
-        Perm<dim+1> perm(unsigned simp, unsigned facet) const;
+        Perm<dim+1> perm(size_t simp, int facet) const;
         /**
          * Deprecated routine to return the gluing permutation
          * associated with the given simplex facet.
@@ -307,8 +307,7 @@ class GluingPerms : public Output<GluingPerms<dim>> {
          * \deprecated This routine has been renamed to perm().  See perm()
          * for details on the preconditions, parameters and return value.
          */
-        [[deprecated]] Perm<dim+1> gluingPerm(unsigned simp, unsigned facet)
-            const;
+        [[deprecated]] Perm<dim+1> gluingPerm(size_t simp, int facet) const;
 
         /**
          * Returns the index into array Perm<dim+1>::Sn_1 describing how the
@@ -357,7 +356,7 @@ class GluingPerms : public Output<GluingPerms<dim>> {
          * investigation (between 0 and \a dim inclusive).
          * @return a reference to the corresponding array index.
          */
-        Index permIndex(unsigned simp, unsigned facet) const;
+        Index permIndex(size_t simp, int facet) const;
 
         /**
          * Offers write access to the index into array Perm<dim+1>::Sn_1
@@ -416,7 +415,7 @@ class GluingPerms : public Output<GluingPerms<dim>> {
          * investigation (between 0 and \a dim inclusive).
          * @return a reference to the corresponding array index.
          */
-        Index& permIndex(unsigned simp, unsigned facet);
+        Index& permIndex(size_t simp, int facet);
 
         /**
          * Returns the triangulation modelled by this set of gluing
@@ -533,7 +532,7 @@ class GluingPerms : public Output<GluingPerms<dim>> {
          * given gluing permutation; this will be between 0 and \a dim!-1
          * inclusive.
          */
-        Index gluingToIndex(unsigned simp, unsigned facet,
+        Index gluingToIndex(size_t simp, int facet,
             const Perm<dim+1>& gluing) const;
 
         /**
@@ -591,8 +590,7 @@ class GluingPerms : public Output<GluingPerms<dim>> {
          * @return the gluing permutation corresponding to the given
          * index into Perm<dim+1>::Sn_1.
          */
-        Perm<dim+1> indexToGluing(unsigned simp, unsigned facet, Index index)
-            const;
+        Perm<dim+1> indexToGluing(size_t simp, int facet, Index index) const;
 
         /**
          * Determines if this and the given gluing permutation set are
@@ -737,7 +735,7 @@ inline void GluingPerms<dim>::swap(GluingPerms& other) noexcept {
 }
 
 template <int dim>
-inline unsigned GluingPerms<dim>::size() const {
+inline size_t GluingPerms<dim>::size() const {
     return pairing_.size();
 }
 
@@ -763,13 +761,13 @@ inline Perm<dim+1> GluingPerms<dim>::gluingPerm(
 }
 
 template <int dim>
-inline Perm<dim+1> GluingPerms<dim>::perm(unsigned simp, unsigned facet) const {
+inline Perm<dim+1> GluingPerms<dim>::perm(size_t simp, int facet) const {
     return indexToGluing(simp, facet, permIndex(simp, facet));
 }
 
 template <int dim>
 inline Perm<dim+1> GluingPerms<dim>::gluingPerm(
-        unsigned simp, unsigned facet) const {
+        size_t simp, int facet) const {
     return indexToGluing(simp, facet, permIndex(simp, facet));
 }
 
@@ -781,7 +779,7 @@ inline typename GluingPerms<dim>::Index& GluingPerms<dim>::permIndex(
 
 template <int dim>
 inline typename GluingPerms<dim>::Index& GluingPerms<dim>::permIndex(
-        unsigned simp, unsigned facet) {
+        size_t simp, int facet) {
     return permIndices_[(dim + 1) * simp + facet];
 }
 
@@ -793,7 +791,7 @@ inline typename GluingPerms<dim>::Index GluingPerms<dim>::permIndex(
 
 template <int dim>
 inline typename GluingPerms<dim>::Index GluingPerms<dim>::permIndex(
-        unsigned simp, unsigned facet) const {
+        size_t simp, int facet) const {
     return permIndices_[(dim + 1) * simp + facet];
 }
 
@@ -813,7 +811,7 @@ inline Perm<dim+1> GluingPerms<dim>::indexToGluing(
 
 template <int dim>
 inline Perm<dim+1> GluingPerms<dim>::indexToGluing(
-        unsigned simp, unsigned facet, Index index) const {
+        size_t simp, int facet, Index index) const {
     if constexpr (standardDim(dim)) {
         return Perm<dim+1>(pairing_.dest(simp, facet).facet, dim) *
             Perm<dim+1>::Sn_1[index] * Perm<dim+1>(facet, dim);
